pyproject.toml
[tool.poetry]
name = "quaternion"
version = "0.1.0"
description = ""
authors = ["Richard G. Edwards <richardgedwards@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy-quaternion = "^2020.7.30"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
